+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - added a widget maker `sgl::Widgets::make<Type>(args)`
 - adding basic actor in `sgl::GameObjects::Actor`
+- adding a templated ressource manager class in `sgl::GameObjects::Manager<T>`
 
 ### Changed
 - the whole widget system is now using the scene graph through the `sgl::Graphics::Node` class